refactor: replace base types
