+++ src/app/providers/StoreProvider/config/services/charactersApi.ts
@@ -2,11 +2,7 @@ import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 
 import { baseUrl } from 'app/constants';
 
-import { Character } from 'pages/home/models';
-
-type ResponseApi = {
-  results: Character[];
-};
+import { Character, ResponseApi } from 'app/types/API';
 
 export const charactersApi = createApi({
   reducerPath: 'charactersApi',
