+++ src/pages/forms/modules/ConfirmMessage/index.ts
@@ -1,0 +1,1 @@
+export { ConfirmMessage } from './ConfirmMessage';
