+++ src/pages/forms/modules/FormCard/index.ts
@@ -1,0 +1,1 @@
+export { FormCard } from "./FormCard";
