+++ src/pages/forms/modules/FormCardList/index.ts
@@ -1,0 +1,1 @@
+export { FormCardList } from "./FormCardList";
